refactor(backend): extract direct emergency fallbacks
- Move direct-node emergency search, synthetic tool SSE events, and salvage recovery helpers into a focused module.\n- Keep the direct runtime call sites and private compatibility names intact.\n- Verify provider-error fallback behavior with targeted tests and ruff.
